First version of AI. Work in progress
AI can be used by sending image of letter in path/api/test/ (this will be changed later). You will receive most likely result as return. AI itself seems to work but it seems to be very opionated about input which is sent to it -> investigating potential reason. Code requires clean up. AI is run each time img is sent to it which is fine at this stage but needs to be fixed later on.